Fixed file reading bug and added write back to disk option
If the last line of the record had no carriage return or line feed the
program would not store it in the data arrays. When the database was
written back out, this record would be missing. The end of file marker
is now checked for as a record terminator.

Added a method to write the data back to disk if changes were made. For
safety, the old file is renamed before writing a new one out.